requirements.txt
fastapi[standard]==0.115.0
httpx
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
